+++ terraform/data/postgres.tf
@@ -3,7 +3,7 @@ resource "helm_release" "postgresql" {
   name       = "postgresql"
   repository = "oci://registry-1.docker.io/bitnamicharts"
   chart      = "postgresql"
-  version    = "16.5.2"
+  version    = "16.7.15"
   namespace  = "data"
 
   set {